@@ -7,10 +7,10 @@ description: A new Flutter project.
 # Both the version and the builder number may be overridden in flutter
 # build by specifying --build-name and --build-number, respectively.
 # Read more about versioning at semver.org.
-version: 2.0.0
+version: 2.2.0
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -18,48 +18,35 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  badges: ^2.0.3
   bloc_provider: ^1.0.0
   cupertino_icons: ^1.0.5
-  cloud_firestore: ^3.1.9
-  google_mobile_ads: ^1.0.0
-  firebase_analytics: ^9.1.1
-  firebase_core: ^1.12.0
-  firebase_crashlytics: ^2.5.2
-  firebase_auth: ^3.3.8
-  firebase_messaging: ^11.2.7
-  firebase_storage: ^10.2.8
+  cloud_firestore: ^4.9.2
+  google_mobile_ads: ^3.0.0
+  firebase_analytics: ^10.5.0
+  firebase_auth: ^4.10.0
+  firebase_core: ^2.16.0
+  firebase_crashlytics: ^3.3.6
+  firebase_messaging: ^14.6.8
+  firebase_storage: ^11.2.7
   flutter_localizations:
     sdk: flutter
   flutter_sticky_header: ^0.6.5
   http: ^0.13.3
-  google_sign_in: ^5.1.0
-  intl: ^0.17.0
+  google_sign_in: ^6.1.5
+  intl: any
   open_file_safe: ^3.2.3
   package_info: ^2.0.2
   path_provider: ^2.0.4
   rxdart: ^0.27.7
   shared_preferences: ^2.0.7
   url_launcher: ^6.0.10
-  collection: ^1.15.0-nullsafety.4
+  collection: ^1.17.0
 
-dependency_overrides:
-  #intl: ^0.17.0
-  firebase_messaging_platform_interface: 3.1.6
-  firebase_storage_platform_interface: 4.0.14
-  firebase_crashlytics_platform_interface: 3.1.13
-  cloud_functions_platform_interface: 5.0.21
-  cloud_firestore_platform_interface: 5.4.13
-  firebase_auth_platform_interface: 6.1.11
-  firebase_database_platform_interface: 0.2.0+5
-  firebase_performance_platform_interface: 0.1.0+4
-  firebase_analytics_platform_interface: 3.0.5
-  firebase_remote_config_platform_interface: 1.0.5
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_launcher_icons: ^0.9.3
+  flutter_launcher_icons: ^0.11.0
 
   test: ^1.16.5
 
