@@ -5,34 +5,26 @@ import 'package:flutter/material.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'app.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
-import 'package:flutter/rendering.dart';
 import 'package:firebase_core/firebase_core.dart';
 
 void main() async {
-  // add this, and it should be the first line in main method
-  WidgetsFlutterBinding.ensureInitialized();
-  MobileAds.instance.initialize();
-  await Firebase.initializeApp();
-  // debugPaintSizeEnabled = true;
-  FlutterError.onError = (FlutterErrorDetails details) {
-    print("FlutterError onerror handler");
-    print(isInDebugMode);
-    if (isInDebugMode) {
-      // In development mode simply print to console.
-      FlutterError.dumpErrorToConsole(details);
-    } else {
-      // In production mode report to the application zone to report to
-      // Crashlytics.
-      Zone.current.handleUncaughtError(details.exception, details.stack!);
-    }
-  };
-
   // FirebaseCrashlytics.instance.enableInDevMode = true;
-  runZoned<Future<Null>>(() async {
+  runZonedGuarded<Future<Null>>(() async {
+    // add this, and it should be the first line in main method
+    WidgetsFlutterBinding.ensureInitialized();
+    MobileAds.instance.initialize();
+    await Firebase.initializeApp();
+    // debugPaintSizeEnabled = true;
+
+    FlutterError.onError = (errorDetails) {
+      FlutterError.dumpErrorToConsole(errorDetails);
+      FirebaseCrashlytics.instance.recordFlutterFatalError(errorDetails);
+    };
+
     runApp(AppProvider(
       child: AppRootWidget(),
     ));
-  }, onError: (dynamic error, dynamic stackTrace) async {
+  }, (Object error, StackTrace stackTrace) async {
     print("runZoned onerror handler");
     if (isInDebugMode) {
       print(error);
